+++ lib/cocoapods-binary-cache/pod-binary/helper/prebuild_sandbox.rb
@@ -3,13 +3,13 @@ require_relative "names"
 module Pod
   class PrebuildSandbox < Sandbox
     # [String] standard_sandbox_path
-    def self.from_standard_sanbox_path(path)
+    def self.from_standard_sandbox_path(path)
       prebuild_sandbox_path = Pathname.new(path).realpath + ".." + PodPrebuild.config.prebuild_sandbox_path
       new(prebuild_sandbox_path)
     end
 
     def self.from_standard_sandbox(sandbox)
-      from_standard_sanbox_path(sandbox.root)
+      from_standard_sandbox_path(sandbox.root)
     end
 
     def standard_sanbox_path