+++ lib/cocoapods-binary-cache/pod-binary/helper/sandbox.rb
@@ -1,0 +1,9 @@
+require_relative "prebuild_sandbox"
+
+module Pod
+  class Sandbox
+    def prebuild_sandbox
+      @prebuild_sandbox ||= Pod::PrebuildSandbox.from_standard_sandbox(self)
+    end
+  end
+end